nuxt: rewrite server handler to use @auth/core
