--- apps/playground-nuxt/.eslintignore
@@ -1,4 +1,0 @@
-dist
-node_modules
-tsconfig.json
-package.json
