--- apps/playground-nuxt/.nuxtrc
@@ -1,1 +1,0 @@
-imports.autoImport=false
