+++ apps/playground-nuxt/app.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <Header />
+    <!-- <Header /> -->
     <NuxtPage />
     <Footer />
   </div>
