--- apps/playground-nuxt/handler.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/runtime/server/handler'
