+++ apps/playground-nuxt/lib/auth/client/index.ts
