+++ apps/playground-nuxt/lib/auth/utils/logger.ts
