+++ apps/playground-nuxt/nuxt.config.ts
@@ -1,9 +1,4 @@
-import MyModule from '../src/module'
-
 export default defineNuxtConfig({
-  modules: [
-    MyModule
-  ],
   // https://v3.nuxtjs.org/migration/runtime-config#runtime-config
   runtimeConfig: {
     secret: process.env.NEXTAUTH_SECRET,
@@ -16,5 +11,12 @@ export default defineNuxtConfig({
   // Fix for GithubProvider is not a function error in Vite
   alias: {
     'next-auth/providers/github': 'node_modules/next-auth/providers/github.js'
+  },
+  vite: {
+    define: {
+      'process.env.NEXTAUTH_URL': JSON.stringify(process.env.NEXTAUTH_URL),
+      'process.env.AUTH_TRUST_HOST': JSON.stringify(process.env.AUTH_TRUST_HOST),
+      'process.env.VERCEL_URL': JSON.stringify(process.env.VERCEL_URL),
+    }
   }
 })
