+++ apps/playground-nuxt/pages/index.vue
