--- apps/playground-nuxt/playground/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "playground",
-  "private": true
-}
