--- apps/playground-nuxt/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  // https://v3.nuxtjs.org/concepts/typescript
-  "extends": "./playground/.nuxt/tsconfig.json"
-}
